adjusted spots slice of state no reviews showing up at this time
